+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.9.0-beta] - 26-09-2025
 
 Migrated to a pure express-server from the previous next.js app.
+
+### Added
+
+- User delete route.